+++ results
@@ -1,4 +1,4 @@
-............F
+............F...............................                             [100%]
 =================================== FAILURES ===================================
 _ TestConverterWorkflowWiring.test_survey_converter_uses_structured_value_offset_editor _
 
@@ -64,5 +64,4 @@ E       AssertionError: 'return surveyValueOffsetEditorController.renderTaskValu
 tests/test_converter_workflow_wiring.py:462: AssertionError
 =========================== short test summary info ============================
 FAILED tests/test_converter_workflow_wiring.py::TestConverterWorkflowWiring::test_survey_converter_uses_structured_value_offset_editor
-!!!!!!!!!!!!!!!!!!!!!!!!!! stopping after 1 failures !!!!!!!!!!!!!!!!!!!!!!!!!!!
-1 failed, 12 passed in 0.03s
+1 failed, 43 passed in 0.35s
